# filter: optimize framerate shaper filter