Reduce complexity of framerate shaper filter. 
The gamma_lut lookup table is accessed for each pixel data for computing sse values. On profiling and analysing, this LUT access turns out to be the most expensive part of the code. If input buffer to motion metric is downscaled (approximated) to quarter of the original frame resolution, the number of LUT access also reduces, thereby reducing the overall computation. 
Before optimization, Framerate Shaper module took 12 seconds. After optimization, it reduced to 2s.
 
Bitrate and quality is not impacted by this optimization.   
 
Results:
Mac M2
BigBuckBunny 4k 60fps - very fast 720p30 x264 encoder
before: 246.59 fps, 1194.83 bitrate
after: 353.19 fps, 1202.98 bitrate
 Very Fast 720p30 vt_h265 encoder
before: 242.14 fps, 392.62 bitrate	
after: 394.77 fps, 391.11 bitrate	
 
Intel Core 7 x86
BigBuckBunny 4k 60fps - Very Fast 720p30 x264 encoder
before: 128.43 fps, 1193.61 bitrate	
after: 206.00 fps, 1201.35 bitrate	
 

Very Fast 720p30 qsv_h265 encoder
before:  130.10 fps, 1653.89 bitrate	
after:  223.56 fps, 1649.02 bitrate	
 
 
Snapdragon Xelite ARM64
BigBuckBunny 4k 60fps - very fast 720p30 x264 encoder
before: 213.93 fps, 1193.26 bitrate	
after: 340.14 fps, 1202.96 bitrate	
 

Very Fast 720p30 mf_h265 encoder
before: 218.06 fps, 236.09 bitrate	
after: 287.51 fps, 234.59 bitrate



**Tested on:**

- [x] Windows 10+  (via MinGW)
- [x] macOS 10.13+
- [ ] Ubuntu Linux