@galad87 2025-02-21
It seems a good idea, however the framerate shaper filter is used to decide which frame to drop when a lower fps is set, so we need to check if it still drops the same ones.

Maybe @jstebbins still have some of the sample that were originally used to test this filter.