@galad87 2025-02-25
Can you keep the existing implementation, and add a flag to hb_motion_metric_init() to enable or disable the approximated one? Maybe we could use the approximated algorithm only for resolution > 1080p or something like this.