@dashsantosh-mcw 2025-06-13
Hi @galad87, thanks for the patch. The PR is rebased on top of that and the functionality is verified to be working.
Also added 10bit neon code for motion metric (from #6907). Could you please review at your earliest convenience?