@galad87 2025-06-15
The metrics are still wrong. The width and height are divided by 4 in the approximated function, but approximate_frame_data() can only rescale to half. That means the approximated function is calculating the metric only on one quarter of the frame.